Add naim.hooks.delete(chain, func).
Change /pyload to import files into a named module instead of __main__. /pyload pynaim/examples/purrbot.py now imports into a module called "purrbot" (based on the filename). An optional second argument can be used to specify the name of the module to import into. These modules can not be reloaded using /pyeval reload(purrbot), but /pyloading a module the same name of an existing module gives the same effect.

When importing a module with /pyload, if that module already exists and has a function called __exit__, call it before wiping the existing module's dict.

Add __exit__ routines to purrbot and weather to deregister their recvfrom hooks.